Conflicts:
	core/src/main/java/com/metsci/glimpse/axis/painter/NumericAxisPainter.java
	core/src/main/java/com/metsci/glimpse/axis/painter/TimeAxisPainter.java
	core/src/main/java/com/metsci/glimpse/axis/painter/TimeXAxisPainter.java
	core/src/main/java/com/metsci/glimpse/axis/painter/TimeYAxisPainter.java

+++ core/src/main/java/com/metsci/glimpse/axis/painter/ColorRightYAxisPainter.java
@@ -68,13 +68,13 @@ public class ColorRightYAxisPainter extends ColorYAxisPainter
     }
 
     @Override
-    public int getColorBarMinX( int width )
+    public float getColorBarMinX( int width )
     {
         return 0;
     }
 
     @Override
-    public int getColorBarMaxX( int width )
+    public float getColorBarMaxX( int width )
     {
         return colorBarSize;
     }
